refactor: move OpenCode skill frontmatter code from loop/ to harness/opencode/
Move ensure_skill_frontmatter(), normalize_project_skill_frontmatter(),
and _normalize_skill_description() from src/loop/helpers.py to
src/harness/opencode/skill_utils.py where they belong.

These functions are OpenCode-specific (YAML frontmatter for skill
discovery) and were the only harness-aware code in the loop/ folder.
Moving them keeps loop/ harness-agnostic — when adding new harnesses
(Goose, OpenHands, Codex), loop/ won't need changes for skill
post-processing.
