+++ src/harness/opencode/__init__.py
@@ -2,5 +2,12 @@
 
 from .options import build_opencode_options
 from .executor import execute_query, parse_response
+from .skill_utils import normalize_project_skill_frontmatter, ensure_skill_frontmatter
 
-__all__ = ["build_opencode_options", "execute_query", "parse_response"]
+__all__ = [
+    "build_opencode_options",
+    "execute_query",
+    "parse_response",
+    "normalize_project_skill_frontmatter",
+    "ensure_skill_frontmatter",
+]
